#156 Renamed components-global.js file to ssr-components.js - removed element component imports and registrations
